refactor: implement with factory design pattern
